pyproject.toml
[project]
name = "curriculum-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair==6.0.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "attrs==25.4.0",
    "blinker==1.9.0",
    "cachetools==6.2.4",
    "certifi==2026.1.4",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorama==0.4.6",
    "distro==1.9.0",
    "dotenv==0.9.9",
    "filelock==3.25.2",
    "fsspec==2026.3.0",
    "gitdb==4.0.12",
    "gitpython==3.1.46",
    "greenlet==3.5.0",
    "groq==1.2.0",
    "h11==0.16.0",
    "hf-xet==1.4.3",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub==1.9.1",
    "idna==3.11",
    "jinja2==3.1.6",
    "jiter==0.12.0",
    "joblib==1.5.3",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "lxml==6.0.2",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "narwhals==2.15.0",
    "networkx==3.6.1",
    "numpy==2.4.1",
    "openai==2.15.0",
    "packaging==25.0",
    "pandas==2.3.3",
    "pillow==12.1.0",
    "playwright==1.58.0",
    "protobuf==6.33.4",
    "pyarrow==23.0.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydeck==0.9.1",
    "pyee==13.0.1",
    "pygments==2.20.0",
    "pypdf2==3.0.1",
    "python-dateutil==2.9.0.post0",
    "python-docx==1.2.0",
    "python-dotenv==1.2.1",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "redis==7.4.0",
    "referencing==0.37.0",
    "regex==2026.4.4",
    "requests==2.32.5",
    "rich==14.3.3",
    "rpds-py==0.30.0",
    "safetensors==0.7.0",
    "scikit-learn==1.8.0",
    "scipy==1.17.1",
    "sentence-transformers==5.3.0",
    "setuptools==81.0.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "streamlit==1.53.0",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tokenizers==0.22.2",
    "toml==0.10.2",
    "torch==2.11.0",
    "torchvision==0.26.0",
    "tornado==6.5.4",
    "tqdm==4.67.1",
    "transformers==5.5.0",
    "typer==0.24.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "urllib3==2.6.3",
    "watchdog==6.0.0",
]
